@@ -1,17 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.2.4" />
-    <PackageReference Include="System.Reactive.Interfaces" Version="4.1.5" />
-    <PackageReference Include="System.Reactive.Linq" Version="4.1.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Tiktack.Web.DataLayer\Tiktack.Web.DataLayer.csproj" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.0.0-preview5.19227.1" />
+    <PackageReference Include="System.Reactive.Interfaces" Version="4.2.0-preview.102" />
+    <PackageReference Include="System.Reactive.Linq" Version="4.2.0-preview.102" />
   </ItemGroup>
 
 </Project>
