@@ -1,9 +1,8 @@
-﻿using Microsoft.AspNet.OData.Extensions;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Sinks.Elasticsearch;
@@ -36,36 +35,36 @@ namespace Tiktack.Web.Api
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
+            services.AddControllers();
             new Bootstrapper().Configure(services, Configuration);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
-            using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
-            {
-                var context = serviceScope.ServiceProvider.GetRequiredService<BaseDbContext>();
-                context.Database.EnsureCreated();
-            }
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
             else
             {
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
-
-            loggerFactory.AddSerilog();
-            app.UseHealthChecks("/healthcheck");
-            app.UseHttpsRedirection();
-            app.UseMvc(op =>
+            using (var serviceScope = app.ApplicationServices.GetService<IServiceScopeFactory>().CreateScope())
             {
-                op.EnableDependencyInjection();
-                op.Expand().MaxTop(100).Select().Filter().OrderBy().Count();
+                var context = serviceScope.ServiceProvider.GetRequiredService<BaseDbContext>();
+                context.Database.EnsureCreated();
+            }
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseEndpoints(endpoint =>
+            {
+                endpoint.MapControllers();
+                endpoint.MapHealthChecks("/healthcheck");
             });
+            loggerFactory.AddSerilog();
         }
+
     }
 }
